requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.5
django-browser-reload==1.11.0
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
